move speaking functions out of face landmark - ad sklearn
